+++ syndicate/connection/__init__.py
@@ -32,6 +32,7 @@ from syndicate.connection.dynamo_connection import DynamoConnection
 from syndicate.connection.ec2_connection import EC2Connection
 from syndicate.connection.elastic_beanstalk_connection import (
     BeanstalkConnection)
+from syndicate.connection.firehose_connection import FirehoseConnection
 from syndicate.connection.iam_connection import IAMConnection
 from syndicate.connection.kinesis_connection import KinesisConnection
 from syndicate.connection.kms_connection import KMSConnection
@@ -157,6 +158,13 @@ class ConnectionProvider(object):
         if region:
             credentials['region'] = region
         return KinesisConnection(**credentials)
+
+    @lru_cache(maxsize=None)
+    def firehose(self, region=None):
+        credentials = self.credentials.copy()
+        if region:
+            credentials['region'] = region
+        return FirehoseConnection(**credentials)
 
     @lru_cache(maxsize=None)
     def application_autoscaling(self, region=None):