+++ syndicate/core/helper.py
@@ -579,3 +579,17 @@ def check_lambda_state_consistency(objected_lambdas: list,
     from syndicate.core.groups import RUNTIME
     return next((True for each in objected_lambdas if subjected_lambdas.get(
         each, {}).get(RUNTIME) == runtime), False)
+
+
+def delete_none(_dict):
+    """Delete None values recursively"""
+    if isinstance(_dict, dict):
+        for key, value in list(_dict.items()):
+            if isinstance(value, (list, dict, tuple, set)):
+                _dict[key] = delete_none(value)
+            elif value is None or key is None:
+                del _dict[key]
+
+    elif isinstance(_dict, (list, set, tuple)):
+        _dict = type(_dict)(delete_none(item) for item in _dict if item is not None)
+    return _dict