+++ syndicate/core/resources/iam_resource.py
@@ -13,6 +13,8 @@
     See the License for the specific language governing permissions and
     limitations under the License.
 """
+import re
+from typing import Optional
 from botocore.exceptions import ClientError
 
 from syndicate.commons.log_helper import get_logger
@@ -320,3 +322,16 @@ class IamResource(BaseResource):
                                                    content=policy_content)
         _LOG.info(f'Updated IAM policy {name}')
         return self.describe_policy(name=name, meta=meta)
+
+    def build_role_arn(self, maybe_arn: str) -> Optional[str]:
+        if not isinstance(maybe_arn, str):
+            return
+        if self.is_role_arn(maybe_arn):
+            return maybe_arn
+        return f'arn:aws:iam::{self.account_id}:' \
+               f'role/{maybe_arn}'
+
+    @staticmethod
+    def is_role_arn(maybe_arn: str) -> bool:
+        return bool(re.match(r'^arn:aws:iam::\d{12}:role/[A-Za-z0-9_-]+$',
+                             maybe_arn))